Gardening: Make a new root dir in the repo called 'tools'.
This directory is going to contain all the tools that we write and use
for *developing*.

Start off by moving /boards/lib/bom-scripts to /tools/bom.

Create /tools/wrapper.  This will later contain a wrapper script that
can be used to invoke all the other scripts in the tools directory.

Create /tools/{pcb,fwman} too.


git-svn-id: https://www.studentrobotics.org/svn/tools@2360 98f67ba8-b320-0410-bda0-b753e294843c