power 0910 + DB: SR codify XBee.
I've created a new category of SR codes: assemblies.  This is for
things that exist as a single component in schematics, but are
physically composed of several components.  The description field for
assemblies must contain a list of the SR part codes it is made from,
in the following syntax:

N(sr-THING) + M(sr-THING2) + P(sr-THING3). DESC

Where N, M and P are integers.  DESC is the assembly description.  The
full-stop before DESC marks the end of the SR codes that make up the
assembly.


git-svn-id: https://www.studentrobotics.org/svn/tools@3058 98f67ba8-b320-0410-bda0-b753e294843c