tools: export_eagle_bom: Avoid race condition when killing eagle.
The export_eagle_bom script gets eagle to run a script that exports a
BOM.  User scripts don't have a method for quitting eagle, so
export_eagle_bom kills eagle when it's created the file.
Unfortunately, it sometimes kills it before it's finished outputting
the whole thing.

This patch makes the script fed to eagle output two BOM files.  We
wait until the second has been created, kill eagle, and then delete
that second file.  We know that the first file is complete as eagle
creates the second file after the first is written :-)

Mmmmm.  Hacky.



git-svn-id: https://www.studentrobotics.org/svn/tools@3093 98f67ba8-b320-0410-bda0-b753e294843c