Swap 4-pin FPC connectors around in library.
Turns out that RS mislabelled these and got top and bottom contacts
the wrong way around.  We're going to assume that RS is moderately
sane and will maintain a part number to physical device mapping.
Obviously this is insanity and will result in some misery later, but
there's not much we can do about it.