Fix farnell screen scraping script
They added some divs inside the availability div so it is no longer
possible to check for the end of the avail div by looking for '</div>'.

The available quantity is now contained in a table so the 'td's should
only be processed when inside the price div.

The available quantity is now contained in an 'a' tag not a 'p'