Add 2x4 way header and socket for beagleboard expansion header
Unfortunately 28 way headers and sockets are few and far between. Two
14 way ones would do but they're still quite expensive and we either
need low-profile ones or a socket which allows the pins to go through
PCB. I managed to find these 8 way ones which are the correct height
and one at each end of the expansion header will provide access to all
of the signals we need.