Replace BSS84LT1 p-channel FET with in stock one and add SOT323 p-channel FET
The BSS84LT1 FET is not in stock at farnell and it seems it won't
be for a while. The BSH202 that I've replaced it with has a lower Vds
of 30V (down from 50V) but a higher Id of 520mA.

I've also added a SOT323 p-channel FET as those SOT-23s are just way
too big ;) It has a Vds of 20V and Id of 300mA. It'll do for almost
all circumstances except for where there's some voltage doubling going
on like the charge pump on the power board.