Change sr-fet-p-bss84lt1g to yet another equivalent
Now the BSH202 is no longer stocked. The IRLML9303 is over-spec'd but has the
right footprint. This FET needs to be able to handle up to 30V as it's used
on the power board to switch the gate of the motor rail FET which can have up
to 28V on it.