Add command to create new instances of assets in the inventory
The command takes a single argument that is the name of an asset template
to create the new asset from. Templates are stored in the ".meta/parts" dir
in the inventory git repository. A unique asset code is generated for the new
asset. This asset code is included in both the name of the file and also in
the file itself to aid movement tracking.