Add an option to 'inv-new-group' to create all assets in an assembly
When initially adding items to the inventory it may be useful to create
all of the assets in an assembly in one go. The '--all' option causes
all of the assets listed in an array called 'elements' in the assembly
template file to be created in the newly created assembly dir.

This commit also adds a '--editor' argument. It performs the same
function as it does in the inv-new-asset command. When creating an
assembly plus the assets inside it this option will cause the editor
to open for each individual asset.