budget: Add diff-to-BudgetTree converter
This patch adds a function that filters and converts a list of
diff-objects into a BudgetTree.  The returned BudgetTree contains the
things that have been added or incremented.

This tree can then be passed to other things that would normally
accept a BudgetTree -- such as the scripts that map the budget into
something presentable to sponsors etc.