Report unknown stock levels correctly
`sr stockcheck` was reporting some RS items as out of stock that were
actually in stock.  This is because we don't know the RS stock level,
and it was incorrectly reporting this as out of stock.  This patch
makes it report it as something that we can't check the stock of.