Make match(...) match on lists and match_single on one node
match(nodes) now filters nodes for those elements that return True
when run through the corresponding match_single(node) function.

This change was made as it makes search tree (although I guess it
wouldn't be much of a search tree then) nodes able to return nodes
different to those passed in.  This could be useful, for example,
as a way of returning the parent of matching parts.