Rejuggle AST so NonTerminals must implement match() themselves
Terminals must implement single_match() which can then be used
by a default match() for Terminals that uses filter().  match()
must take a list of parts and return a (possible empty) list of
parts.  AND, OR, NOT will filter the output lists accordingly,
but generalising this way allows functions (comming soon) to
return different parts to the ones passed in.