budget-check: Remove 10% fudge from closed lines
budget-check's role up to now has been to ensure that the budgeted
amount is less than the amounts listed in funds-in.yaml.  The values
in funds-in.yaml are a factor of 1.1 times lower than the amount of
money that actually came in.  This was so that budget lines could have
10% more than their estimated value spent against them.

We have recently introduced the concept of a closed budget line.  A
closed budget line is one that cannot be spent against any more, and
their 'cost' field is set to the amount actually spent.  These budget
lines therefore no longer need a 10% fudge factor anymore.

This patch modifies budget-check so that the fudge factor is removed
from closed budget lines when calculating the total amount of money
that we need.

An example:

Let's say that our budget contains three budget lines: a, b, and
c.  Let's also say that we've received £T in total from our sponsor.
budget-check's role is to check that:

	1.1(a + b + c) < T

The comparison it actually makes is:

	a + b + c < T/1.1

The sum of the lines in funds-in.yaml is T/1.1

Let's say we now close line a.  line a no longer has any fudge factor,
so now we need to ensure that:

	a + 1.1(b + c) < T

We rearrange this slightly to accommodate for the aforementioned
funds-in.yaml being T/1.1:

	a/1.1 + b + c < T/1.1

Change-Id: I51ad48709e10143e7e708c577770efbc019e08df