Mark the output arg required.
It's used without checking when writing the resulting file, so it's
already required, this just marks it as such. Also add an alternative
name which makes the error message a little clearer.

It might be nice at some point to do work out what name the user
wants based on the input, but for now this is the quicker route to
get it working.

Change-Id: I372a5e1c8905f11371c763481b9688c0ca6fc5e7