Add repo-checking and root-finding to spending.py
In general, the spending tools want to know where the root of
spending.git is.  They also want to be sure that they are actually
looking at spending.git, not some other git repository.  At the
moment, the various tools use a bit of a hacky approach to this.  This
commit adds a more robust version of this functionality to
spending.py.  The new function spending.find_root():

1) Checks that the repository is indeed spending.git by checking for
   the presence of the first commit in spending.git.
2) Returns the root directory of the repository.