Add a config loader to the sr python lib
A bunch of our tools often prompt the user for things like a username
and password.  It becomes tiresome typing one's username in all the
time, so it'd be good if one could set this in a config file and not
be bothered ever again.  So this patch introduces a config file that
is aimed at doing exactly that.

The library looks for config in two places:

 1) ~/.sr/tools/config.yaml -- Not for user editing
 2) ~/.sr/config.yaml -- For user editing

The values found in the latter will override those found in the
former.

So, to set one's SR username, one would write the following into
~/.sr/config.yaml:

user: beedogs