Optionally store one's SR password in a keyring
If one sets "use_keyring" in the config to True, then the
get_password() function of a Config object will attempt to lookup the
user's password in the keyring.

At the time of writing, the keyring library
( https://pypi.python.org/pypi/keyring ) supports a variety of
keyrings from different platforms (Gnome, KDE, OSX, and some kind of
Windows thing).

After the first password entry, the password is stored in the keyring,
and the user should never have to enter their password again.