budget-check: Add check against spending.git
We want to make sure that budget changes don't invalidate references
that spending.git entries make.  This patch adds an option to
`budget-check` that takes the path of a spending.git clone.  If
provided it will check that all of the budget lines referred to from
that clone do in fact exist.