budget: Parse line costs using python
The cost field was previously being evaluated using sympy.  Sympy is
the source of most of the slow-down in parsing the budget at the
moment.  This patch completely removes the use of sympy, and uses
python itself to evaluate the cost.

Since we want numbers to be exact, the cost string is mangled using
the python parser.  All float literals in the python are extended to
be decimals *before* they are interpreted as floats.

The other result of this is that we can now properly round costs to
the nearest penny.  This is good, as it means that when I do budget
wrangling I don't have to continuously hack around some rounding
errors :)

This takes running budget-check down to 0.470s for me.