Fix adding multiple dependencies to a ticket.
Work out in advance which dependencies are being added and which are
already present. Use this information (rather than local looping) to
work out whether or not to do anything.

In particular, this fixes the presumed typo where trying to add more
than one dependency at a time would fail because the entire list was
added the first time around.

Also, in the case where some of the dependencies are already present,
there is no need to bail immediately, so instead print warnings about
those already present and add the rest.

Change-Id: Ie7a1409931ee17dae87cda25b2a04ccacdd07a7f