Allow optional "uncertainty" field to budget lines
We are in a situation where we are extremely close to spending some
money on some things.  We are much more certain about the levels of
certainty involved in our estimates of the spending amount, because
we're looking at the quotes that we actually want to act on.  We also
need to free up some budget to be able to do this.

So this patch introduces an optional "uncertainty" field to budget
lines.  By default this is 0.1, which is the fudge factor we were
applying previously.  When a line is closed, this is forced to 0.

This patch updates budget-check and budget.py to support this new
field.