Update YAML and move to loading most inputs as untrusted
This introduces use of `safe_load` almost everywhere that we load
YAML data. The few places that we don't do so are more complex
(and in areas of the code which are unused and pending removal).

We also declare support for PyYAML v5 (but not v4 as that version
broke things), which has this safe API and now warns about use of
plain `yaml.load`.