Install Sphinx upfront in CI
This is a bit of a hack though it fixes CI for now. It's not clear
why letting setuptools install Sphinx as part of the `install_requires`
dependencies doesn't work, however as we're likely going to need
to rework how installs work soon anyway (since python packaging
as a whole is shifting) it doesn't seem worth spending too much
time on fixing it right now.